feat: Fixture 데이터 따로 빼기.
- api/__fixture__/ 로 임시 mock 데이터 모으는 공간 구성.
